main: don't limit MAXPROCS to 1 - memory fragmentation seems to be fixed
I had seen crazy memory usage w/ maxprocs > 1 before, I think as a
result of the fan-out of requests to different threads, and each
thread doing a bunch of short-lived allocations.  This seems fixed in
tip/1.5, so remove any call to set the maxprocs, where it will default
to 1 for Go 1.4 and earlier, and the number of CPUs in Go1.5 and
later.